Moron. I was editing the wrong file!!!
